backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
pydantic==2.7.1
python-dotenv==1.0.1
